Fix build error, consolidate dis6 and util libraries
Fix build error from a fresh clone/install - The build script for
examples used find_package() to try and find the opendis library built
by this project. This works when the project has already been built
and installed once, but does not work on a fresh installation. This
replaces find_package() with direct use of the openDIS library built
by the project.

Consolidate dis6 and util libraries - the premake version of OpenDIS
puts the symbols from dis6/* and utils/* in the same library, the
cmake version now does the same.

This also removes the EXPORT directive under install() - this was
giving me errors on my machine and the documentation wasn't being very
helpful as to how to fix it